some no-op changes
we are trying to get into a state that runs
Ready: False   kustomize build failed: kustomization.yaml is empty

This error since ~Flux 2.2 (Kustomize 5.3) was resolved in Flux 2.2.3

It can still be triggered by adding nothing to a kustomization.yaml, or
fixed up by removing it. It should not be empty or we cannot build it
successfully. Flux generates an error when the file is empty, it is not
a ready Kustomization and can't proceed to dependencies without content.

Signed-off-by: Kingdon P Barrett <kingdon@tuesdaystudios.com>